Refactor Moving Averages: Extract Standard Moving Averages to new file
